+++ services/preview/servicefabric/mgmt/2018-02-01-preview/servicefabric/applicationtypes.go
